+++ cli/cli/command_str_consts/command_str_consts.go
@@ -25,6 +25,7 @@ const (
 	FilesCmdStr           = "files"
 		FilesUploadCmdStr = "upload"
 		FilesStoreWebCmdStr = "storeweb"
+		FilesStoreServiceCmdStr = "storeservice"
 	ModuleCmdStr = "module"
 		ModuleExecCmdStr = "exec"
 	ServiceCmdStr = "service"