+++ cli/docs/changelog.md
@@ -1,7 +1,7 @@
 # TBD
 ### Features
-* Updated `files_artifact_mounting` internal test in order to test duplicate mountpoints
 * Added `files storeservice` to store files from a service in the enclave
+* Updated `files_artifact_mounting` internal test in order to test duplicate mountpoints
 
 ### Changes
 * Upgraded `kurtosis core lib` to 1.43.6 which adds `EnclaveContext.StoreFilesFromService`